pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "groq>=1.1.2",
    "openai>=2.32.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
]
